+++ src/lib/menu/MenubarItem.svelte
@@ -1,10 +1,19 @@
 <script lang="ts">
 	import { Menubar } from 'bits-ui';
+
+	export let text: string;
+	export let key: string | undefined = undefined;
 </script>
 
 <Menubar.Item
-	class="mx-1 cursor-default rounded-md py-1 pl-4 pr-12 text-left text-slate-400 hover:bg-gray-700 hover:text-slate-200"
+	class="group mx-1 flex cursor-default items-center justify-between rounded-md py-1 pl-4 text-left hover:bg-gray-700 {key
+		? 'pr-4'
+		: 'pr-10'}"
 	on:click
 >
-	<slot />
+	<span class="text-slate-400 group-hover:text-slate-300">{text}</span>
+
+	{#if key}
+		<span class="pl-6 font-mono text-sm text-slate-500 group-hover:text-slate-400">{key}</span>
+	{/if}
 </Menubar.Item>