+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Option to filter out enabled mods
+
+### Removed
+
+- Home page with changelog. The default when starting the app is now the profile page instead.
+
+### Changed
+
+- Minor UI tweaks and fixes
+
+### Fixed
+
+- Last updated showing NaN on Linux
+- Sometimes being able to scroll the whole window
+
 ## 0.8.5 (2024-09-11)
 
 ### Added